requirements.txt
streamlit
pandas
sklearn.decomposition
sklearn.preprocessing
matplotlib.pyplot
seaborn
numpy